% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_sample_size.R
\name{standard_sample_size}
\alias{standard_sample_size}
\title{Standard sample size for a two-arm continuous outcome trial}
\usage{
standard_sample_size(sd, effect, power = 0.9, alpha = 0.05, allocation = 1)
}
\arguments{
\item{sd}{Population standard deviation.}

\item{effect}{Expected treatment difference.}

\item{power}{Desired statistical power.}

\item{alpha}{Type I error rate.}

\item{allocation}{Allocation ratio of experimental to control arm.}
}
\value{
Required sample size per treatment arm.
}
\description{
Calculates the required main-trial sample size per treatment arm
assuming the population standard deviation is known.
}
\examples{
standard_sample_size(
  sd = 1,
  effect = 0.5,
  power = 0.90,
  alpha = 0.05
)

}
