% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucl_sample_size.R
\name{ucl_sample_size}
\alias{ucl_sample_size}
\title{Main trial sample size using the upper confidence limit method}
\usage{
ucl_sample_size(
  pilot_n,
  sd,
  effect,
  power = 0.9,
  alpha = 0.05,
  allocation = 1,
  conf_level = 0.8
)
}
\arguments{
\item{pilot_n}{Pilot trial sample size per treatment arm.}

\item{sd}{Pilot estimate of the standard deviation.}

\item{effect}{Expected treatment difference.}

\item{power}{Desired statistical power.}

\item{alpha}{Type I error rate.}

\item{allocation}{Allocation ratio of experimental to control arm.}

\item{conf_level}{Confidence level for the upper confidence limit.}
}
\value{
A list containing the degrees of freedom, upper confidence
  limit for the variance, upper confidence limit for the standard
  deviation, and main-trial sample size per treatment arm.
}
\description{
Calculates the required main-trial sample size per treatment
arm using the upper confidence limit (UCL) approach described
by Browne.
}
\details{
Browne recommends an 80% upper confidence
level. However, Sim and Lewis, set X at 0.95 or
the 95% level. @details See also \code{ssPilot::optimized_ucl_sample_size}.
}
\examples{
ucl_sample_size(
  pilot_n = 16,
  sd = 1,
  effect = 0.50,
  power = 0.90,
  alpha = 0.05,
  conf_level = 0.80
)

}
\references{
Browne RH. On the use of a pilot sample for sample size
determination. Stat Med 1995; 14: 1933–1940.

Sim J and Lewis M. The size of a pilot study for a clinical
 should be calculated in relation to considerations of
precision and efficiency. J Clin Epidemiol 2012; 65:
301–308.
}
