% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_sample_size.R
\name{nct_sample_size}
\alias{nct_sample_size}
\title{Main trial sample size using the non-central t-distribution}
\usage{
nct_sample_size(
  pilot_n,
  sd,
  effect,
  power = 0.9,
  alpha = 0.05,
  allocation = 1,
  max_n = 10000
)
}
\arguments{
\item{pilot_n}{Pilot trial sample size per treatment arm.}

\item{sd}{Pilot estimate of the standard deviation.}

\item{effect}{Expected treatment difference.}

\item{power}{Desired statistical power.}

\item{alpha}{Type I error rate.}

\item{allocation}{Allocation ratio of experimental to control arm.}

\item{max_n}{Maximum main-trial sample size per arm to search.}
}
\value{
A list containing the pilot-trial sample size per arm,
  pilot-trial degrees of freedom, required main-trial sample size
  per arm, main-trial degrees of freedom, central t critical value,
  non-central t quantile, and the right-hand side of Equation (4).
}
\description{
Calculates the required main-trial sample size per treatment
arm using the non-central t-distribution (NCT) approach
described by Julious and Owen.
}
\examples{
nct_sample_size(
  pilot_n = 12,
  sd = 1,
  effect = 0.50,
  power = 0.90,
  alpha = 0.05
)

}
\references{
Julious SA and Owen RJ. Sample size calculations for
clinical studies allowing for uncertainty about the variance.
Pharmaceut Stat 2006; 5: 29–37.
}
