% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nct_sample_size.R
\name{nct_sample_size}
\alias{nct_sample_size}
\title{Main trial sample size using the non-central t-distribution}
\usage{
nct_sample_size(
  pilot_n,
  sd,
  effect,
  power = 0.9,
  alpha = 0.05,
  allocation = 1,
  max_n = 10000
)
}
\arguments{
\item{pilot_n}{Pilot trial sample size per treatment arm.}

\item{sd}{Pilot estimate of the standard deviation.}

\item{effect}{Expected treatment difference.}

\item{power}{Desired statistical power.}

\item{alpha}{Type I error rate.}

\item{allocation}{Allocation ratio of experimental to control arm.}

\item{max_n}{Maximum main-trial sample size per arm to search.}
}
\value{
A list containing the pilot-trial sample size per arm,
  pilot-trial degrees of freedom, required main-trial sample size
  per arm, main-trial degrees of freedom, central t critical value,
  non-central t quantile, and the right-hand side of Equation (4).
}
\description{
Calculates the required main-trial sample size per treatment
arm using the non-central t-distribution (NCT) approach
described by Whitehead et al. (2016).
}
\examples{
nct_sample_size(
  pilot_n = 12,
  sd = 1,
  effect = 0.50,
  power = 0.90,
  alpha = 0.05
)

}
\references{
Whitehead, A. L., Julious, S. A., Cooper, C. L., & Campbell, M. J.
(2016). Estimating the sample size for a pilot randomised trial
to minimise the overall trial sample size for the external pilot
and main trial for a continuous outcome variable.
Statistical Methods in Medical Research, 25(3), 1057-1073.
}
