% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_ucl_sample_size.R
\name{optimized_ucl_sample_size}
\alias{optimized_ucl_sample_size}
\title{Optimized pilot sample size using the UCL method}
\usage{
optimized_ucl_sample_size(
  sd,
  effect,
  power = 0.9,
  alpha = 0.05,
  allocation = 1,
  conf_level = 0.8,
  min_pilot = 10,
  max_pilot = 100
)
}
\arguments{
\item{sd}{Anticipated standard deviation based on prior knowledge,
previous studies, published literature, or other available evidence.}

\item{effect}{Expected treatment difference.}

\item{power}{Desired statistical power for the main trial.}

\item{alpha}{Type I error rate.}

\item{allocation}{Allocation ratio of experimental to control arm.}

\item{conf_level}{Confidence level for the upper confidence limit.}

\item{min_pilot}{Minimum pilot sample size per treatment arm.}

\item{max_pilot}{Maximum pilot sample size per treatment arm to search.}
}
\value{
A list containing the optimal pilot sample size per arm,
  the corresponding main-trial sample size per arm, the total
  sample size per arm, the confidence level used, and a data frame
  containing the optimization results for all candidate pilot
  sample sizes.
}
\description{
Finds the pilot-trial sample size per treatment arm that minimizes
the combined pilot and main-trial sample size using the upper
confidence limit approach.
}
\examples{
optimized_ucl_sample_size(
  sd = 1,
  effect = 0.50,
  power = 0.90,
  alpha = 0.05,
  conf_level = 0.80
)

}
\references{
Whitehead, A. L., Julious, S. A., Cooper, C. L., & Campbell, M. J. (2016).
 Estimating the sample size for a pilot randomised trial to minimise the overall trial sample size for the external pilot and main trial for a continuous outcome variable.
 Statistical Methods in Medical Research, 25(3), 1057–1073.
}
